devoir/requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0